feat(application,product,search) : implement external search using meili search engine service and create  rabbit mq publisher service to synchronize  data between sql and meili search
- Create ISearchService
- Crerate MeiliSearchService Implement ISearchService
- Create IRabbitMqPublisherService
- Create RabbitMqPublisherService Implement IRabbitMqPublisherService  With PublishProductCreatedEventAsync Function To Publish ProductCreatedEvent

+++ ECommerce.Application/Carts/DTOs/CartDto.cs
@@ -3,7 +3,7 @@
     public sealed record CartItemDto(
         int ProductId,
         int Quantity,
-        int Price
+        decimal Price
     );
     public sealed record CartDto(
         int Id,
